requirements.txt
anthropic
openai
mistralai
google-genai
python-dotenv